TestProject2/requirements.txt
certifi==2022.12.7
charset-normalizer==3.1.0
click==8.1.3
colorama==0.4.6
databricks-cli==0.17.6
idna==3.4
lxml==4.9.2
oauthlib==3.2.2
Pillow==9.5.0
PyJWT==2.6.0
python-pptx==0.6.21
requests==2.28.2
six==1.16.0
tabulate==0.9.0
urllib3==1.26.15
XlsxWriter==3.0.9
